+++ lib/src/duration_extension.dart
@@ -4,8 +4,6 @@ extension DurationExtension on Duration? {
   /// Get localized duration.
   ///
   /// This uses `Intl.DateFormat`.
-  ///
-  /// Creates a new DateFormat, using the format specified by [pattern].
   String format({
     String? locale,
     String pattern = 'HH:mm:ss',
@@ -15,11 +13,15 @@ extension DurationExtension on Duration? {
       return '';
     }
 
-    final DateFormat dateFormat =
-        DateFormat(pattern, locale ?? Platform.localeName);
+    final DateFormat dateFormat = DateFormat(
+      pattern,
+      locale ?? Platform.localeName,
+    );
 
-    final DateTime dateTime =
-        DateTime.fromMillisecondsSinceEpoch(this!.inMilliseconds, isUtc: isUtc);
+    final DateTime dateTime = DateTime.fromMillisecondsSinceEpoch(
+      this!.inMilliseconds,
+      isUtc: isUtc,
+    );
 
     return dateFormat.format(dateTime);
   }