+++ test/date_extension_test.dart
@@ -1,19 +1,24 @@
-// ignore_for_file: avoid_print
-
 import 'package:dart_intl_extensionz/dart_intl_extensionz.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:test/test.dart';
 
 void main() {
   initializeDateFormatting();
-  test('format', () {
+  group('format', () {
     final DateTime value = DateTime(2020, 6, 29);
-    final String result1 = value.format();
-    final String result2 = value.format(locale: 'en_GB');
-    final String result3 = value.format(locale: 'de_DE');
+    test('en_US', () {
+      final String result = value.format(locale: 'en_US');
+      expect(result, '6/29/2020');
+    });
 
-    expect(result1, '6/29/2020');
-    expect(result2, '29/06/2020');
-    expect(result3, '29.6.2020');
+    test('en_GB', () {
+      final String result = value.format(locale: 'en_GB');
+      expect(result, '29/06/2020');
+    });
+
+    test('de_DE', () {
+      final String result = value.format(locale: 'de_DE');
+      expect(result, '29.6.2020');
+    });
   });
 }