+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Dart Intl Extensionz
 
+## 0.0.2
+
+- Fix description length.
+- Improve comments.
+- Improve `String.toDateTime` default pattern.
+- Improve tests.
+
 ## 0.0.1
 
 - Initial dev preview.